% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{callfest}
\alias{callfest}
\alias{callfest-package}
\title{callfest.}
\usage{
callfest(..., FUN = list, N = 1)
}
\arguments{
\item{...}{List of arguments which will be used for combinations.}

\item{FUN}{The function to be applied to combinations.}

\item{N}{The number of function invokations to each of combinations.}
}
\value{
An instance of class \code{callfest} which encapsulates
results and the elapsed time.
}
\description{
callfest.

Calls a function over a list of arguments which are combinations
of the provided dotted arguments.
}

