% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.callfest}
\alias{print.callfest}
\title{Pretty prints the \code{callfest} instance}
\usage{
\method{print}{callfest}(x, ...)
}
\arguments{
\item{x}{A \code{callfest} instance to be pretty printed.}

\item{...}{Additional arguments to print method.}
}
\description{
Pretty prints the \code{callfest} instance
}

