% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{callfest}
\alias{callfest}
\alias{callfest-package}
\title{callfest.}
\usage{
callfest(control, ...)
}
\arguments{
\item{control}{A \code{callfestControl} object of a function to be
applied to combinations.}

\item{...}{List of arguments which will be used for combinations.}
}
\value{
An instance of class \code{callfest} which encapsulates
results and the elapsed time.
}
\description{
callfest.

Calls a function over a list of arguments which are combinations
of the provided dotted arguments.
}
\examples{
## Create a control instance:
control <- callfestControl(list, N=1, parallel=FALSE)

## Call fest:
result <- callfest(control, a=1:2, b=c("a", "b"))

## This is equivalent to:
result <- callfest(list, a=1:2, b=c("a", "b"))
}

